+++ src/features/authentication/model/otp/sending/verify-otp.test.ts
@@ -2,11 +2,11 @@ import { checkOneTimePassword } from '@shared/api'
 import type { RegisterResult } from '@shared/types'
 import { beforeEach, describe, expect, it, vi } from 'vitest'
 
-import { monitorInformation, MONITORING_EVENTS } from '../../monitoring/monitoring'
+import { monitorInformation, MONITORING_EVENTS } from '../../../monitoring/monitoring'
 import { verifyOtp } from './verify-otp'
 
 // mocks
-vi.mock('../../monitoring/monitoring', () => ({
+vi.mock('../../../monitoring/monitoring', () => ({
   monitorInformation: vi.fn(),
   MONITORING_EVENTS: {
     AUTH_REGISTER_SUBMITTED: 'AUTH_REGISTER_SUBMITTED',
