+++ src/features/settings/index.ts
@@ -1,4 +1,4 @@
 export { translationKeys } from './i18n/keys'
-export { getInitialLanguage } from './model/language-service'
-export { getInitialTheme } from './model/theme-service'
-export { getThemeFromStorage, setThemeToStorage } from './model/theme-storage'
+export { getInitialLanguage } from './model/set_language/language-service'
+export { getInitialTheme } from './model/set_theme/theme-service'
+export { getThemeFromStorage, setThemeToStorage } from './model/set_theme/theme-storage'
