+++ src/features/settings/model/set_language/language-storage.test.ts
@@ -1,7 +1,7 @@
 import { getStorageItem, LanguageCode, setStorageItem } from '@shared/lib'
 import { beforeEach, describe, expect, it, vi } from 'vitest'
 
-import { monitorInformation } from '../monitoring/monitoring'
+import { monitorInformation } from '../../monitoring/monitoring'
 import { getLanguageFromStorage, setLanguageToStorage } from './language-storage'
 
 vi.mock('@shared/lib', () => ({
@@ -9,7 +9,7 @@ vi.mock('@shared/lib', () => ({
   setStorageItem: vi.fn()
 }))
 
-vi.mock('../monitoring/monitoring', () => ({
+vi.mock('../../monitoring/monitoring', () => ({
   monitorInformation: vi.fn(),
   MONITORING_EVENTS: {
     SETTINGS_LANG_WRITE: 'SETTINGS_LANG_WRITE',
